requirements.txt
Flask==3.1.0
numpy==2.2.3
pandas==2.2.3
PyYAML==6.0.2
PyYAML==6.0.2
torch==2.5.1
